requirements.txt
# Core trading libraries
ccxt>=4.0.0
ccxt-pro>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
numba>=0.57.0

# Technical analysis
TA-Lib>=0.4.25
btalib>=1.0.0
vectorbt>=0.25.0

# Machine learning (optional)
scikit-learn>=1.3.0

# API and web framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
websockets>=11.0.0

# Database
psycopg2-binary>=2.9.0
redis>=4.6.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0
sentry-sdk>=1.29.0

# Configuration and secrets
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Utilities
requests>=2.31.0
aiohttp>=3.8.0
python-telegram-bot>=20.4.0
